GAdManager: Refactor `AmpAd` component
